+++ packages/stream_chat_flutter/lib/src/utils/utils.dart
@@ -1,4 +1,5 @@
 export 'device_segmentation.dart';
 export 'extensions.dart';
 export 'helpers.dart';
+export 'jitter.dart';
 export 'typedefs.dart';