+++ packages/stream_chat_flutter/CHANGELOG.md
@@ -3,6 +3,8 @@
 🐞 Fixed
 
 - Fixed `.replaceMentions` not escaping special characters in the username.
+- Fixed `GradientAvatars` for users with same-length IDs would have identical
+  colors. [[#2369]](https://github.com/GetStream/stream-chat-flutter/issues/2369)
 
 ## 9.16.0